Change "up to X energy units" for extensions.
As of right now, docs say that extensions can contain up to 50 energy units. However, that is not the case with 7th and 8th RCL, as at that point they upgrade and can store more energy (100 and 200 respectively). My suggestion is to change that paragraph to say more accurate information.